Separate routes from apis
