+++ views/login.ejs
@@ -10,7 +10,7 @@
     Currently logged in user: <%- currentUser %>
     <br>
     Login
-    <form action="/login" method="post">
+    <form action="/api/login" method="post">
       <input name="username" type="text" placeholder="Username" />
       <input name="email" type="email" placeholder="Email" />
       <input name="password" type="password" placeholder="Password" />
