+++ views/register.ejs
@@ -8,7 +8,7 @@
   </head>
   <body>
     Register
-    <form action="/register" method="post">
+    <form action="/api/register" method="post">
       <input name="username" type="text" placeholder="Username" />
       <input name="email" type="email" placeholder="Email" />
       <input name="password" type="password" placeholder="Password" />
